chore(date-input): move to packages
