+++ packages/components/inputs/date-input/src/date-input-open.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { DateInput } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const value = '2018-11-13';
