+++ packages/components/inputs/date-input/src/date-input.js
@@ -19,11 +19,11 @@ import {
   parseInputToDate,
   getIsDateInRange,
 } from '@commercetools-uikit/calendar-utils';
-import CalendarBody from '../../internals/calendar-body';
-import CalendarMenu from '../../internals/calendar-menu';
-import CalendarHeader from '../../internals/calendar-header';
-import CalendarContent from '../../internals/calendar-content';
-import CalendarDay from '../../internals/calendar-day';
+import CalendarBody from '../../../../../src/components/internals/calendar-body';
+import CalendarMenu from '../../../../../src/components/internals/calendar-menu';
+import CalendarHeader from '../../../../../src/components/internals/calendar-header';
+import CalendarContent from '../../../../../src/components/internals/calendar-content';
+import CalendarDay from '../../../../../src/components/internals/calendar-day';
 import messages from './messages';
 
 const DateInput = (props) => {
