+++ packages/components/inputs/date-input/src/date-input.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { render, fireEvent } from '../../../../test/test-utils';
+import { render, fireEvent } from '../../../../../test/test-utils';
 import DateInput from './date-input';
 
 // This component is used to enable easy testing.
