+++ packages/components/inputs/date-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './date-input';
+
+export { default as version } from './version';
