--- src/components/inputs/date-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './date-input';
